requirements.txt
torch
adam_atan2
adam-atan2
einops
tqdm
coolname
pydantic
argdantic
wandb
omegaconf
hydra-core
huggingface_hub
packaging 
ninja 
wheel 
setuptools 
setuptools-scm
pydantic-core
huggingface_hub 
numba
triton
# Image Captioning (HRM/FuseLIP/Swin)
ftfy
regex
peft
transformers
accelerate
pycocoevalcap
pycocotools
torchvision
pillow